Add bower and node deps to repo
Preivously, bower and node dependencies were defined
in configuration files, but the actual libraries
were ignored from the git repo. This has the
advantages of:

* Reducing the size of the repo
* Keeping the git history clean of changes to
  individual vendor library files

However, this also means that every time we run a
build or deployment, Travis has to download the node
and bower dependencies. On a few occasions, certain
npm or bower repos have failed to serve the files we
need. Then our build and deployment breaks, even
though the library itself is passing the build