Refactor Animation objects
BREAKING CHANGE

- Use base layers as a "master" controller
  for all other layer frames.
  So instead of setting the opacity on the animation,
  we can simply set the layer's opacity.
  We can also use layer.setMap(null) to remove it.

This better decouples the Animation from the layer. Previously,
if a layer was being animated, the client could not manipulate
the base layer without messing up the animation. Now, the client
can use the base layer the same as before it was animated, and
all animation frames will follow suit.