TileAnimation: do not automatically preload layers
Preloaded has a couple of significant downsides
- Can cause application to freeze up when adding
  multiple layers (in Safari, this was causing
  application to crash with 3 layers)
- Additional bandwidth costs for Aeris API servers

This commit completely disables preloading. However, the 'preloadLayer_'
method is left intact, if we want to provide some type of
interface for preloading layers in the future.

If you are considering preloading, I would try to think of some way
to prevent setting too many layers to the map at once -- I believe
this is what is causing the performance issues. Instead, maybe
set layers to map at a setInterval