Prevent aeris.Collection from passing unwanted options to models
Consider:

  new aeris.Collection([{ model: 'A' }], { strategy: null })

our intent here is to create a collection with no strategy.
However, all options to the Collection get passed on to any models
created by the Collection constructor. So in this case, our model
would be created with a null strategy. This commit fixes that, so only
the standard collection options are passed through.