Fix: animation bug for tiles with no future times
Was throwing an error when attempting to go to a
future time, when the layer had no future times available.

This fix also improves transisions, when tiles are are not loaded.
When transitioning to a not-yet-loaded tile layer, the animation
will instead show the closest available layer in the same tense.
This prevents the empty animations you may see when running through
an animation loop for the first time.