Refactor TileAnimation, fix setTo/setFrom
was not re-loading layers on bounds change

removed layerLoader, tileLayerFactory
- not really needed, now that we're just grabbing static times

I deleted all test coverage, and for that I am truly sorry.
But the test coverage was mostly testing implementation details,
and not actual behavior.
And it relied heavily on mocked out helpers (which no longer exist).
Also, test coverage for this time of UI thing is super hard.